# Change UTF-8 decoder to avoid maximum stack size being exceeded