@dchest 2017-09-01
Thank you! It was a silly of me to think that I could get away with `String.fromCharCode.apply`.

I've also added a failing test, which your PR fixes.

Regarding performance and memory use, I suspect that modern JS engines will work better with a simple string concatenation -- at least memory-wise -- but if there's VM which doesn't optimize this, it would behave pretty badly, so I think your solution is a nice middle ground. I'll probably later add support for `TextEncoder` if it's available, so that newer VMs won't even have to worry about running this code.