@dchest 2017-09-01
BTW, to run benchmarks you can do: `yarn build && yarn bench` inside the `packages/utf8` directory. Here are results:

Before:

```
utf8 encode                               12356 ops           0.04 ms/op     24690.57 ops/sec     64.87 MiB/s
utf8 decode                               26780 ops           0.02 ms/op     53506.09 ops/sec    140.58 MiB/s
```

After:

```
utf8 decode                                4960 ops           0.10 ms/op      9904.95 ops/sec     26.02 MiB/s
```

Quite a performance hit, but since the implementation is incorrect, it doesn't matter.

With string concatenation (`let s = ""; s += String.fromCharCode(...)`):

```
utf8 decode                               10346 ops           0.05 ms/op     20680.78 ops/sec     54.34 MiB/s
```

Maybe I'll change it to string concatenation later, but it's okay for now.