@ernieturner 2017-09-01
Thanks for the quick turnaround!